SAFETY MONITORING SYSTEM
safety monitoring system for covid-19 is a model for health safety system with the help of temperature sensors and IR sensors and this project having major advantages and applications in the present situation...
software description: 
The Arduino Integrated Development Environment (IDE) is a cross-platform application (for Windows, macOS, Linux) that is written in the programming language Java. It is used to write and upload programs to Arduino compatible boards, but also, with the help of 3rd party cores, other vendor development boards.[2] 

The source code for the IDE is released under the GNU General Public License, version 2.[3] The Arduino IDE supports the languages C and C++ using special rules of code structuring.[4] The Arduino IDE supplies a software library from the Wiring project, which provides many common input and output procedures. User-written code only requires two basic functions, for starting the sketch and the main program loop, that are compiled and linked with a program stub main() into an executable cyclic executive program with the GNU toolchain, also included with the IDE distribution.[5] The Arduino IDE employs the program avrdude to convert the executable code into a text file in hexadecimal encoding that is loaded into the Arduino board by a loader program in the board's firmware.[6] .